Attempt to work around #161.
Force HTTP protocol in all of the URLs that we generate. If my analysis in
mybinder.org start working again. If it is not, this change should be easy
to revert.

This is not a real solution, but hopefully this will get us working in this
use case until we can finally fix the SSL stuff for real.