lib/wwt.js: defensive coding for multiple views of a single model
In the Jupyter widget system, it is possible to create multiple views of a
single widget model. Views can be "detached", in which case their underlying
DOM elements are removed from the document. This is what happens if you create
a WWT view in a JupyterLab notebook, use the "Create a new view for this
output" context menu item to move the output to a new tab, and then hide the
original output. There is one underlying widget model and two views, only one
of which has active DOM elements.

The wrinkle for pywwt is that all of our widget state is stored inside the
<iframe> that does the rendering, so we're not really a real model-view setup.
The way things currently work, operations that affect the widget state are
passed as message from Python to JS, which are then distributed to all of the
views. So long as all views receive the same messages, things should stay in
sync. (Although if you create a widget, do some things to it, and then create
a new view of it, that view will have missed historical messages and so will
be out of sync.)

The problem I discovered was that if you send messages to a "detached" widget
view, you sometimes get JS exceptions. And apparently the code that
distributes the messages just terminates, preventing the control messages from
being sent to other widget views. This means that those views basically stop
responding to any Python controls.

This is bad news for the workflow of creating a WWT view then moving it to a
separate tab, which is something we very much want to support for UX reasons.
The real solution is to Do The Right Thing when a widget is detached, so that
the JS exceptions don't happen in the first place, but that's going to take
some time to get right. In the meantime, let's just log and swallow those JS
exceptions, so that the message passing can keep on working to the views that
aren't broken. This will keep things limping along.