Try to get Windows CI working again
When I added the hack to get Python 3.6 running on our AppVeyor Windows CI,
everything passed, but now we have a problem. The images generated on
Windows/Py36 are now different from what we've historically expected. This
isn't shocking since my hack was to fix the version of pyqt to avoid a library
incompatibility issue, for Python 3.6 only.

Let's try applying the same version constraint *everywhere*. From my
experiments, this seems to make Windows/Py3.x create images agreeing with the
'webkit' test corpus, while Windows/Py2.7 still agrees with 'webkit_win'. One
clear difference between the two is the reported OpenGL renderer, so revise
the test logic to factor that in when choosing which corpus to compare
against.