webpack.config.js: put embed-bundle.js in the right place
Our `docs` directory doesn't use a `source/` subdirectory, so take that out of
the Webpack config. Add the generated files to .gitignore. Seeing as this file
wasn't landing in the right place and I don't see it referenced anywhere else
in the repo, I suspect that it's not actually doing anything.