docs/index.rst: iterate introductory material and update Binder links
This page still had links to `mybinder.org` for demo notebooks, which don’t
work for the time being because of WWT’s well-known issues with SSL. Update
these to our customized BinderHub instance. Sadly, those don’t work *either*
because we need to put out a new release of pywwt to keep compatibility with
the latest versions of JupyterLab.

Also iterate the introductory material to align with the README more closely,
and be more aggressive about the AAS branding.