RELEASE.rst: tweak the release notes
Mention updating the `stable` branc to update the ReadTheDocs docs. Also
mention creating a new release on GitHub.

Finally, recommend a workflow in which the release commits are submitted via a
pull request, so that that there is a CI run on the release candidate commit
and the HEAD commit that keeps master at a development version.