pywwt/conftest.py: attempt to fix CI
The previous time implementation had a bug where `set_current_time()` was off
by a month, because JavaScript has the *amazing* thing where the constructor
`Date(year, month, day)` uses zero-based month numbers, but 1-based everything
else. This PR uses a different approach that should fix this issue, but that
has the side effect of changing the images generated during the CI process.
Make the problem go away by altering the reference date to be 2017 Feb 1,
which was not what was intended before, but was what was actually being used.