tests: finally fix the MacOS rendering issue?
Some local testing identifiers some behavior that seems likely to be
responsible for the frustratingly inconsistent results from macOS
testing. On macOS, it seems that every so often a Qt mainloop iteration
takes *~~30 seconds* to execute, and we need to execute several such
iterations for the WWT rendering to fully update properly. See more
discussion in `wait_for_test()`.