front: rework multi-view management
Try to straighten out our handling of the fact that Jupyter widgets can
have multiple "views".

There are some new long comments in the source explaining the approach.
We basically have the widget model track a "current" view for reporting
state back. Because reparenting an iframe in the DOM causes it to
reload, we can't do anything clever about preventing the creation of
multiple WWT iframes for a single view without breaking the important
JupyterLab pop-out-a-new-view functionality.