pywwt/nbextension/static/wwt.html: remove trailing comma
It looks like the Qt JavaScript interpreter on macOS doesn't like the
trailing comma in our function call, causing the embedded webpage to
break. If correct this is a little scary in terms of what it means for
our ability to use remotely advanced JS constructs, but let's see if
this keeps things going for now.

Also some general whitespace futzing, but the key change is to the
`startMode` line.