pywwt/interactive_figure: move these files out of nbextension
They don't need to be in the nbextension static directory. In fact we
should probably keep the nbextension static directory as clean as we can
manage since those files are copied into the Jupyter notebook data tree.
But this is more about clarifying the different roles played by
different sets of static data files.