Rearrange webserver static files
The static files served up by `pywwt.jupyter_server` (for our Jupyter
server extension) and `pywwt.data_server` (for the Qt widget) don't need
to be mixed in with the `pywwt.nbextension` JS files implementing the
Jupyter notebook widget extension.

While we're at it, change `wwt.html` to be `widget/index.html`, to lay
the groundwork for being able to replace the simple hand-coded widget
HTML/JS with a more complex collection of files if needed.