Bundle the research app in the pywwt distribution
Here we (ab)use the pywwt JS frontend package to distribute the WWT
research app in the pywwt `web_static` folder. We're not fully wired up
to use it everywhere yet, but this sets the stage for making it
available.

While the JupyterLab integration currently runs the research app from
its `wwtassets.org` origin, there are valid reasons to want to run a
local version. In particular, in the SciServer science platform the WWT
engine was unable to view FITS data because their system has some extra
security layers that block cross-origin requests to its Jupyter server
no matter what we try.

On the other hand, installations running the research app from an
official WWT domain will automatically pick up app upgrades, and if/when
we have user logins with interesting functionality (e.g. "save this
image to my account"), users will be able to take advantage of that
functionality.