Port the ipywidgets frontend to use the research app
Here we start using the research app for the ipywidgets Jupyter
frontend. This requires an extensive rework of the `frontend` JavaScript
package to adapt to the app messaging framework, which in turn required
a bunch of improvements to the app's messaging scheme to support the
ipywidgets use case. In particular, one ipywidgets "widget" might
include several "views", each of which is represented by an app, and the
message-passing infrastructure will receive messages from all active
apps at once. So we need to throw in various unique IDs that allow
everyone to keep track of which messages are meant for them.

This work is not complete. The JupyterLab app interface needs to be
updated to track some of the changes that were needed for the ipywidgets
frontend, and there are plenty of corner cases left dealing with cases
where widget views go away or stop being responsive. On the Python side,
we should pay attention to when there's actually a running app and when
it goes away, and raise exceptions as needed.

All that being said, this ought to get us going here, and I'd like to
get this baseline committed before starting to chase down issues like
the ones mentioned above.