Allow the app to come and go, not just become ready once
For both the Lab app and ipywidgets use cases, we need to continuously
monitor whether the app is available: it might crash, or the user might
do something weird to its iframe, or they might simply close its window.
So we can't just have a one-time "ready" signal: we need events
indicating when the app is responding and when it isn't.

The UX here is somewhat poor: when the app isn't responding, instead
of raising an exception with a wonky stack trace, we should yield
a terse error message telling the user what to do.

Proper handling of coming-and-going of the app also exposes how we need
to have a way for pywwt to fully synchronize its internal state with the
state of the app. Right now, if you're using JupyterLab and connect a
second client to the app, it will reset the foreground/background
imagery. That's not desirable.

We also-also need a way to signal the client that the engine's internal
state has been reset so that references to, e.g. annotations and
data-table layers have been invalidated.