Port the Qt client over to use the research app
This turns out to be very straightforward. Compared to Jupyter, there's
just one client and one app to worry about. The Qt web driver frameworks
don't seem to support sending and receiving web messages, but we work
around that with some manual JavaScript futzing.

The tests will fail here since the research app adds chrome that's not
expected in our reference files. Since glue uses the Qt client and we
don't want to suddenly add a bunch of chrome to its WWT windows, we'll
add some functionality that defaults the Qt clients to hiding the
chrome. That will require us to add some new messages to the research
app.