Rework image-testing approach
For the image-output testing suite, we used to have a library of
reference images for each platform and some complex logic to figure out
which image library to use depending on various OS and OpenGL
parameters. This turned out to be pretty annoying to manage since it
doesn't encourage re-use of reference images: if there's one specific
renderer setup that requires a special reference image for one case, you
have to set up a whole image library for that one setup.

Here we switch to just have a collection of reference images for each
test case. The test suite passes if *any* of these reference images
matches the observed output. This simplifies things a lot and makes
reuse much easier.