pywwt/tests/data/refimg_qt_full_step5/c.png: minor update to make Windows happy
This reference image is just barely failing the test suite, I think because of an
issue relating to our handling of the host system's timezones. I should straighten
that out, but don't feel like working on that right now. The file is tiny, so let's
just go ahead and replace it.