pywwt/layers.py: add support for HiPS catalogs
This is what all of the async work is building up to. In order to
populate HiPS catalogs on the fly, we *really* benefit from the async
framework because we have to roundtrip with the WWT frontend repeatedly.

This significantly builds on the work of @imbasimba in #289, but builds
on the work I've been doing to unify everything with the "research app"
backend and these async foundations.