pywwt/qt.py: bump up the app liveness deadline to 10 seconds
Addresss #309 the straightforward way. Slow operations can hold up the
Qt mainloop enough that the backend convinces itself that the app has
become nonresponsive.

Note that for the ipywidgets and JupyterLab app backends, the liveness
testing is done in the JS frontend code, which really ought to not
suffer from this possibility. So those deadlines are staying at their
current values, for now.