pywwt/jupyter_relay.py: add WWT Kernel Data Relay implementation
This switches the Jupyter implementation to use the new Kernel Data
Relay (KDR) server extension rather than bundled one in
`pywwt/jupyter_server.py`. The key difference is that the KDR scheme is
much more robust, flexible, and extensible than the bundled server
extension, and it makes it possible for us to correctly expose tiled
data to WWT and other clients.