Update reference image corpus for research-app 0.7.x
The new WebGL2 FITS rendering code and imageset zoom setup alters some
of our expected outputs.

Note in particular that in the `image_layer_gal` test, the WebGL1 and
WebGL2 renderers produce fairly different opacity effects. As best I can
tell, they are both behaving as intended, it's just that the shaders
(TileShader and FitsShader) approach opacity in different ways. It might
be possible to harmonize them, but it doesn't seem worth the effort
right now.