frontend: fix URL canonicalization
It helps to make sure that basic function calls are done correctly! This
should fix the widget when the baseUrl is non-empty. Also try to create
nice URLs regardless of whether the baseUrl ends in a slash or not -- I
don't know what we can expect there, and I don't actually know if a
double-slash is a problem, but it doesn't hurt to be careful.