pywwt/jupyter_server.py: add a separate static-file handler
This is to ease migration as we start switching to the WWT Kernel Data
Relay for hosting kernel data files.

There is still a value in having pywwt act as a Jupyter *server*
extension (which it does in addition to being a *notebook* extension and
a *JupyterLab* extension). This is because some Jupyter(Lab) setups deny
all cross-origin requests, which means that we have to host a local copy
of the research app for it to load any data from the kernel. Since pywwt
already bundles the research app, it's a natural place to do that, and
will remain so even if everyone uses the KDR, because it can host that
local copy even if no kernels are running. But to prevent clashes
between older and newer versions, we want to use a separate URL
namespace. So here we add support for serving files out of
`$baseurl/wwtstatic/`. For the time being, this is redundant with the
`/wwt/` code, but that should go away once we switch over to relying on
the KDR.

This change should get the lab app working better on platforms like
SciServer with cross-origin restrictions. There may still be some
issues, though, if the kernels and servers don't share home directories,
due to the way that the non-KDR service code works. At a minimum the KDR
rollout will still be important to enable the service of locally tiled
datasets.