Restructure the API docs
Put the documentation for each module/package into its own file to make
things more tractable. The `api.rst` index page is now just a stub table
of contents, which should be filled out in more detail. More
restructuring to be done but, combined with the new theme, this is
*much* nicer.