pywwt/jupyter_relay.py: make sure to always include seqnums in messages
When we were returning error messages, we weren't including sequence
numbers in the content payload. This used to be fine, but the current
KDR protocol specification is quite clear that there always has to be a
seqnum, and their lack was causing 404 messages to be lost because the
relay was just dropping such messages on the floor.