ci: drop down to only one Mac build
The unreliability of the Mac builds makes the CI/CD process pretty
fragile. Let's use the Linux builds to test compatibility across Python
versions, and just do one Mac build to reduce the chances of random
failure.

Of course, I'd like to fix the Mac builds to work more reliably, but
it's a challenging debugging cycle with the CI systems, and in a few
past tries I haven't been able to make any progress on making the
timeouts go away.